Moved all interfaces under shared resources project.
