+++ EQM-GQE.DATA/QuestionnaireContext.cs
@@ -1,3 +1,4 @@
+using EQM_GQE.SHARED_RESOURCES.Interfaces;
 using EQM_GQE.SHARED_RESOURCES.Models;
 using Microsoft.EntityFrameworkCore;
