+++ EQM-GQE.LOGICAL/QuestionnaireLogic.cs
@@ -4,7 +4,7 @@ using System.Collections.Generic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
-using EQM_GQE.DATA.Repositories;
+using EQM_GQE.SHARED_RESOURCES.Interfaces;
 
 namespace EQM_GQE.LOGICAL
 {
